+++ src/language/rule/rule.ts
@@ -86,8 +86,12 @@ export interface IRuleMetadata {
 
 export type RuleType = "functionality" | "maintainability" | "style" | "typescript";
 
+export type RuleSeverity = "warning" | "error";
+
 export interface IOptions {
     ruleArguments: any[];
+    ruleSeverity: RuleSeverity;
+
     ruleName: string;
     disabledIntervals: IDisabledInterval[];
 }
@@ -109,6 +113,7 @@ export interface IRuleFailureJson {
     failure: string;
     fix?: Fix;
     name: string;
+    ruleSeverity: string;
     ruleName: string;
     startPosition: IRuleFailurePositionJson;
 }
@@ -227,6 +232,7 @@ export class RuleFailure {
     private startPosition: RuleFailurePosition;
     private endPosition: RuleFailurePosition;
     private rawLines: string;
+    private ruleSeverity: RuleSeverity;
 
     constructor(private sourceFile: ts.SourceFile,
                 start: number,
@@ -239,6 +245,7 @@ export class RuleFailure {
         this.startPosition = this.createFailurePosition(start);
         this.endPosition = this.createFailurePosition(end);
         this.rawLines = sourceFile.text;
+        this.ruleSeverity = "warning";
     }
 
     public getFileName() {
@@ -273,6 +280,14 @@ export class RuleFailure {
         return this.rawLines;
     }
 
+    public getRuleSeverity() {
+        return this.ruleSeverity;
+    }
+
+    public setRuleSeverity(value: RuleSeverity) {
+        this.ruleSeverity = value;
+    }
+
     public toJson(): IRuleFailureJson {
         return {
             endPosition: this.endPosition.toJson(),
@@ -280,6 +295,7 @@ export class RuleFailure {
             fix: this.fix,
             name: this.fileName,
             ruleName: this.ruleName,
+            ruleSeverity: this.ruleSeverity.toUpperCase(),
             startPosition: this.startPosition.toJson(),
         };
     }