+++ src/linter.ts
@@ -34,7 +34,7 @@ import { findFormatter } from "./formatterLoader";
 import { ILinterOptions, LintResult } from "./index";
 import { IFormatter } from "./language/formatter/formatter";
 import { createLanguageService, wrapProgram } from "./language/languageServiceHost";
-import { Fix, IRule, RuleFailure } from "./language/rule/rule";
+import { Fix, IRule, RuleFailure, RuleSeverity } from "./language/rule/rule";
 import { TypedRule } from "./language/rule/typedRule";
 import * as utils from "./language/utils";
 import { loadRules } from "./ruleLoader";
@@ -136,6 +136,18 @@ class Linter {
             }
         }
         this.failures = this.failures.concat(fileFailures);
+
+        // add rule severity to failures
+        const ruleSeverityMap = new Map(enabledRules.map((rule) => {
+            return [rule.getOptions().ruleName, rule.getOptions().ruleSeverity] as [string, RuleSeverity];
+        }));
+        for (const failure of this.failures) {
+            const severity = ruleSeverityMap.get(failure.getRuleName());
+            if (severity === undefined) {
+                throw new Error(`Severity for rule '${failure.getRuleName()} not found`);
+            }
+            failure.setRuleSeverity(severity);
+        }
     }
 
     public getResult(): LintResult {
@@ -152,12 +164,14 @@ class Linter {
 
         const output = formatter.format(this.failures, this.fixes);
 
+        const errorCount = this.failures.filter((failure) => failure.getRuleSeverity() === "error").length;
         return {
-            failureCount: this.failures.length,
+            errorCount,
             failures: this.failures,
             fixes: this.fixes,
             format: formatterName,
             output,
+            warningCount: this.failures.length - errorCount,
         };
     }